Split out PSF3D into a separate file
