+++ gammapy/irf/__init__.py
@@ -4,6 +4,7 @@
 from .effective_area import *
 from .psf_core import *
 from .psf_table import *
+from .psf_3d import *
 from .psf_analytical import *
 from .psf_king import *
 from .energy_dispersion import *
